+++ src/types.ts
@@ -17,37 +17,6 @@ export interface ExtraFieldsCSV {
   hour: string
 }
 
-export interface JsonFromSbs {
-  messageType: string
-  transmissionType: string
-  sessionID: string
-  aircraftID: string
-  icao24: string | number
-  flightID: string
-  dateMessageGenerated: string | number
-  timeMessageGenerated: string | number
-  dateMessageLogged: string | number
-  timeMessageLogged: string | number
-  callsign: string
-  geoaltitude: string
-  groundspeed: string
-  track: string
-  latitude: string
-  longitude: string
-  vertical_rate: string
-  squawk: string
-  alert: string
-  emergency: string
-  spi: string
-  onground: string
-  haveLabel?: string
-  label?: string
-  altitude?: string
-  last_position?: string
-  lastcontact?: string
-  hour?: string
-}
-
 export type JsonMessage = Record<string, string | boolean | number | undefined>
 
 export type CsvRow = {